refactor: restructure entities and update imports for album, artist, track, favorites, and user modules
