+++ README.md
@@ -1,9 +1,13 @@
 # Home Library Service
 
+## Description
+
+The Home Library Service is a music library management system that allows you to manage your music collection. You can work with artists, albums, tracks, and create your favorites list.
+
 ## Prerequisites
 
-- Git - [Download & Install Git](https://git-scm.com/downloads).
-- Node.js - [Download & Install Node.js](https://nodejs.org/en/download/) and the npm package manager.
+- Git - [Download & Install Git](https://git-scm.com/downloads)
+- Node.js - [Download & Install Node.js](https://nodejs.org/en/download/) and the npm package manager
 
 ## Downloading
 
@@ -24,35 +28,21 @@ npm start
 ```
 
 After starting the app on port (4000 as default) you can open
-in your browser OpenAPI documentation by typing http://localhost:4000/doc/.
-For more information about OpenAPI/Swagger please visit https://swagger.io/.
+in your browser OpenAPI documentation by typing http://localhost:4000/api/
+For more information about OpenAPI/Swagger please visit https://swagger.io/
 
 ## Testing
 
 After application running open new terminal and enter:
 
-To run all tests without authorization
-
+To run all tests:
 ```
 npm run test
 ```
 
-To run only one of all test suites
-
+To run specific test:
 ```
-npm run test -- <path to suite>
-```
-
-To run all test with authorization
-
-```
-npm run test:auth
-```
-
-To run only specific test suite with authorization
-
-```
-npm run test:auth -- <path to suite>
+npm run test -- <path-to-test-file>
 ```
 
 ### Auto-fix and format
@@ -61,12 +51,63 @@ npm run test:auth -- <path to suite>
 npm run lint
 ```
 
-```
-npm run format
-```
+## Usage
 
-### Debugging in VSCode
+The service provides the following functionality:
 
-Press <kbd>F5</kbd> to debug.
+### Users
+- Get all users
+- Get single user by id
+- Create user
+- Update user's password
+- Delete user
 
-For more information, visit: https://code.visualstudio.com/docs/editor/debugging
+### Artists
+- Get all artists
+- Get single artist by id
+- Create artist
+- Update artist
+- Delete artist
+
+### Albums
+- Get all albums
+- Get single album by id
+- Create album
+- Update album
+- Delete album
+
+### Tracks
+- Get all tracks
+- Get single track by id
+- Create track
+- Update track
+- Delete track
+
+### Favorites
+- Get all favorites
+- Add track to favorites
+- Delete track from favorites
+- Add album to favorites
+- Delete album from favorites
+- Add artist to favorites
+- Delete artist from favorites
+
+## Implementation Details
+
+- The service uses in-memory database for data storage
+- All entities have UUID identifiers
+- When an artist, album, or track is deleted:
+  - Their IDs are removed from favorites (if present)
+  - References to them in other entities become null
+- Non-existing entities cannot be added to favorites
+- All requests and responses use application/json format
+- Users' passwords are excluded from responses
+
+## Development
+
+The project is built using:
+- Node.js
+- NestJS
+- TypeScript
+- Class-validator for request validation
+- Swagger/OpenAPI for API documentation
