+++ src/album/dto/create-album.dto.ts
@@ -1,4 +1,4 @@
 import { OmitType } from '@nestjs/mapped-types';
-import { Album } from '../album.entity';
+import { Album } from '../entities/album.entity';
 
 export class CreateAlbumDto extends OmitType(Album, ['id'] as const) {}
