+++ src/artist/dto/create-artist.dto.ts
@@ -1,4 +1,4 @@
 import { OmitType } from '@nestjs/mapped-types';
-import { Artist } from '../artist.entity';
+import { Artist } from '../entities/artist.entity';
 
 export class CreateArtistDto extends OmitType(Artist, ['id'] as const) {}
