+++ src/artist/dto/update-artist.dto.ts
@@ -1,4 +1,4 @@
 import { OmitType } from '@nestjs/mapped-types';
-import { Artist } from '../artist.entity';
+import { Artist } from '../entities/artist.entity';
 
 export class UpdateArtistDto extends OmitType(Artist, ['id'] as const) {}
