+++ src/artist/entities/artist.entity.ts
