+++ src/favorites/entities/favorites.entity.ts
