+++ src/track/dto/create-track.dto.ts
@@ -1,4 +1,4 @@
 import { OmitType } from '@nestjs/mapped-types';
-import { Track } from '../track.entity';
+import { Track } from '../entities/track.entity';
 
 export class CreateTrackDto extends OmitType(Track, ['id'] as const) {}
